re-structured the file structure of FE
