+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This file lists all changes to this project, grouped by versions that follow [Se
 - Add +x permissions to `be/generate_secret.sh`
 - Makefile testing rules to allow running BE and FE together or
   individually.
+- Re-structured the file structure of FE.
 
 ## [0.6.1] - 2020-03-08
